pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "html2text>=2024.2.26",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "streamlit>=1.44.1",
    "trafilatura>=2.0.0",
]
